# docs/openapi.yaml
openapi: 3.0.3
info:
  title: PSRide API
  description: |
    PSRide API — a serverless Node.js backend for ride-sharing among University of Ilorin members.

    ## Authentication
    Most endpoints require a Bearer JWT token obtained via `/auth/login` or `/auth/register`.
    Include it in the `Authorization` header: `Bearer <token>`.

    Tokens expire based on `JWT_EXPIRY` config. Use `/auth/refresh-token` with a refresh token to get a new access token.

    ## Rate Limiting
    All endpoints are rate-limited. Limits vary by endpoint type:
    - General API: 100 req/15 min
    - Auth endpoints: 20 req/15 min
    - Login: 10 req/15 min
    - Password reset: 5 req/15 min
    - OTP: 5 req/15 min
    - Search: 30 req/15 min
    - Booking: 10 req/15 min
    - SOS: 5 req/15 min

    Rate limit headers returned: `X-RateLimit-Limit`, `X-RateLimit-Remaining`, `X-RateLimit-Reset`.

    ## Payment Model
    Phase 1 MVP uses cash-only payments with a 6-digit verification code flow.

    ## Error Handling
    All errors follow a consistent format with error codes from the `ErrorCodes` enum.
  version: 1.3.0
  contact:
    name: PSRide Team
  license:
    name: MIT
servers:
  - url: https://jk4sd35xw3.execute-api.eu-west-1.amazonaws.com/dev/api/v1
    description: Development
  - url: http://localhost:3000/api/v1
    description: Local Development
...
paths:
...
  /uploads/presign:
    post:
      tags: [Uploads]
      summary: Request a presigned upload URL
      description: |
        Returns a presigned S3 PUT URL for direct browser-to-S3 uploads.
        The URL is valid for 15 minutes. After uploading, call `/uploads/confirm` to verify.

        **Upload flow:**
        1. Call this endpoint to get a presigned URL
        2. `PUT` the file to the returned `uploadUrl` with the matching `Content-Type` header
        3. Call `/uploads/confirm` with the returned `key` to verify the upload

        **Categories:**
        - `profile-photo` — max 5 MB
        - `driver-document` — max 10 MB (license, registration, insurance)
        - `vehicle-photo` — max 10 MB (requires `entityId` for vehicle ID)
      operationId: requestUploadUrl
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PresignUploadRequest'
            examples:
              profilePhoto:
                summary: Profile photo
                value:
                  category: profile-photo
                  contentType: image/jpeg
              driverDocument:
                summary: Driver document
                value:
                  category: driver-document
                  contentType: image/png
                  entityId: license-front
              vehiclePhoto:
                summary: Vehicle photo
                value:
                  category: vehicle-photo
                  contentType: image/webp
                  entityId: VEHICLE#abc123
      responses:
        '201':
          description: Presigned upload URL generated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/PresignUploadResponse'
              example:
                success: true
                message: Upload URL generated
                data:
                  uploadUrl: "https://carpool-dev-uploads.s3.eu-west-1.amazonaws.com/uploads/profile-photos/USER123/abc.jpg?X-Amz-Algorithm=..."
                  key: "uploads/profile-photos/USER123/abc.jpg"
                  expiresAt: "2026-03-16T12:15:00.000Z"
                  maxSize: 5242880
                  contentType: image/jpeg
                timestamp: "2026-03-16T12:00:00.000Z"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
  /uploads/confirm:
    post:
      tags: [Uploads]
      summary: Confirm an upload
      description: |
        Verifies that a file was successfully uploaded to S3.
        Call this after uploading to the presigned URL.
        Returns the image metadata including content type and file size.
      operationId: confirmUpload
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UploadKeyRequest'
            example:
              key: "uploads/profile-photos/USER123/abc.jpg"
      responses:
        '200':
          description: Upload confirmed
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ConfirmUploadResponse'
              example:
                success: true
                message: Upload confirmed
                data:
                  key: "uploads/profile-photos/USER123/abc.jpg"
                  imageUrl: "https://carpool-dev-uploads.s3.eu-west-1.amazonaws.com/uploads/profile-photos/USER123/abc.jpg"
                  contentType: image/jpeg
                  size: 245760
                timestamp: "2026-03-16T12:00:00.000Z"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: File not found in S3 (not yet uploaded or expired)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /uploads/view:
    post:
      tags: [Uploads]
      summary: Get a signed view URL
      description: |
        Returns a presigned GET URL for viewing a private image.
        The URL is valid for 1 hour. Use this to display images in the frontend.
      operationId: getViewUrl
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UploadKeyRequest'
            example:
              key: "uploads/profile-photos/USER123/abc.jpg"
      responses:
        '200':
          description: View URL generated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          viewUrl:
                            type: string
                            description: Presigned GET URL (valid 1 hour)
                          expiresIn:
                            type: integer
                            description: URL expiry in seconds
                            example: 3600
        '401':
          $ref: '#/components/responses/Unauthorized'
  /uploads:
    delete:
      tags: [Uploads]
      summary: Delete an upload
      description: Permanently deletes a file from S3. You can only delete your own uploads.
      operationId: deleteUpload
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UploadKeyRequest'
            example:
              key: "uploads/profile-photos/USER123/abc.jpg"
      responses:
        '200':
          description: Upload deleted
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          deleted:
                            type: boolean
                            example: true
                          key:
                            type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT access token obtained from `/auth/login` or `/auth/register`.
        Include as: `Authorization: Bearer <token>`

  # ──────────────────────────────────────────────
  # PARAMETERS
  # ──────────────────────────────────────────────
...
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Operation successful
        data:
          type: object
          nullable: true
        timestamp:
          type: string
          format: date-time
      required: [success, message, timestamp]
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
            message:
              type: string
              example: Validation failed
            details:
              type: object
              nullable: true
        timestamp:
          type: string
          format: date-time
      required: [success, error, timestamp]
...
    PresignUploadRequest:
      type: object
      required: [category, contentType]
      properties:
        category:
          type: string
          enum: [profile-photo, driver-document, vehicle-photo]
          description: Type of image being uploaded
        contentType:
          type: string
          enum: [image/jpeg, image/png, image/webp, image/heic]
          description: MIME type of the image
        entityId:
          type: string
          maxLength: 100
          description: Optional entity identifier (e.g., vehicle ID for vehicle photos)
    UploadKeyRequest:
      type: object
      required: [key]
      properties:
        key:
          type: string
          maxLength: 500
          description: S3 object key returned from the presign endpoint
          example: "uploads/profile-photos/USER123/abc.jpg"
    PresignUploadResponse:
      type: object
      properties:
        uploadUrl:
          type: string
          description: Presigned S3 PUT URL — upload file here via PUT with matching Content-Type
        key:
          type: string
          description: S3 object key — use this for confirm, view, and delete
        expiresAt:
          type: string
          format: date-time
          description: URL expiry timestamp (15 minutes from generation)
        maxSize:
          type: integer
          description: Maximum allowed file size in bytes
          example: 5242880
        contentType:
          type: string
          description: Content type to use in the PUT request
          example: image/jpeg
    ConfirmUploadResponse:
      type: object
      properties:
        key:
          type: string
          description: S3 object key
        imageUrl:
          type: string
          description: Full S3 URL of the uploaded image
        contentType:
          type: string
          description: MIME type of the uploaded file
        size:
          type: integer
          description: File size in bytes

    # ── Error Code Reference ───────────────────
...
  responses:
...
    Unauthorized:
      description: Authentication required or invalid token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: UNAUTHORIZED
              message: Authentication required
            timestamp: "2026-02-27T12:00:00.000Z"
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: FORBIDDEN
              message: You do not have permission to perform this action
            timestamp: "2026-02-27T12:00:00.000Z"
...
    ValidationError:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: VALIDATION_ERROR
              message: Validation failed
              details:
                - field: email
                  message: email must be a valid email address
                  type: string.email
            timestamp: "2026-02-27T12:00:00.000Z"
...